+++ library/common/src/main/java/com/google/android/exoplayer2/MediaMetadata.java
@@ -383,6 +383,108 @@ public final class MediaMetadata implements Bundleable {
       return this;
     }
 
+    /** Populates all the fields from {@code mediaMetadata}, provided they are non-null. */
+    public Builder populate(@Nullable MediaMetadata mediaMetadata) {
+      if (mediaMetadata == null) {
+        return this;
+      }
+      if (mediaMetadata.title != null) {
+        setTitle(mediaMetadata.title);
+      }
+      if (mediaMetadata.artist != null) {
+        setArtist(mediaMetadata.artist);
+      }
+      if (mediaMetadata.albumTitle != null) {
+        setAlbumTitle(mediaMetadata.albumTitle);
+      }
+      if (mediaMetadata.albumArtist != null) {
+        setAlbumArtist(mediaMetadata.albumArtist);
+      }
+      if (mediaMetadata.displayTitle != null) {
+        setDisplayTitle(mediaMetadata.displayTitle);
+      }
+      if (mediaMetadata.subtitle != null) {
+        setSubtitle(mediaMetadata.subtitle);
+      }
+      if (mediaMetadata.description != null) {
+        setDescription(mediaMetadata.description);
+      }
+      if (mediaMetadata.mediaUri != null) {
+        setMediaUri(mediaMetadata.mediaUri);
+      }
+      if (mediaMetadata.userRating != null) {
+        setUserRating(mediaMetadata.userRating);
+      }
+      if (mediaMetadata.overallRating != null) {
+        setOverallRating(mediaMetadata.overallRating);
+      }
+      if (mediaMetadata.artworkData != null) {
+        setArtworkData(mediaMetadata.artworkData, mediaMetadata.artworkDataType);
+      }
+      if (mediaMetadata.artworkUri != null) {
+        setArtworkUri(mediaMetadata.artworkUri);
+      }
+      if (mediaMetadata.trackNumber != null) {
+        setTrackNumber(mediaMetadata.trackNumber);
+      }
+      if (mediaMetadata.totalTrackCount != null) {
+        setTotalTrackCount(mediaMetadata.totalTrackCount);
+      }
+      if (mediaMetadata.folderType != null) {
+        setFolderType(mediaMetadata.folderType);
+      }
+      if (mediaMetadata.isPlayable != null) {
+        setIsPlayable(mediaMetadata.isPlayable);
+      }
+      if (mediaMetadata.year != null) {
+        setRecordingYear(mediaMetadata.year);
+      }
+      if (mediaMetadata.recordingYear != null) {
+        setRecordingYear(mediaMetadata.recordingYear);
+      }
+      if (mediaMetadata.recordingMonth != null) {
+        setRecordingMonth(mediaMetadata.recordingMonth);
+      }
+      if (mediaMetadata.recordingDay != null) {
+        setRecordingDay(mediaMetadata.recordingDay);
+      }
+      if (mediaMetadata.releaseYear != null) {
+        setReleaseYear(mediaMetadata.releaseYear);
+      }
+      if (mediaMetadata.releaseMonth != null) {
+        setReleaseMonth(mediaMetadata.releaseMonth);
+      }
+      if (mediaMetadata.releaseDay != null) {
+        setReleaseDay(mediaMetadata.releaseDay);
+      }
+      if (mediaMetadata.writer != null) {
+        setWriter(mediaMetadata.writer);
+      }
+      if (mediaMetadata.composer != null) {
+        setComposer(mediaMetadata.composer);
+      }
+      if (mediaMetadata.conductor != null) {
+        setConductor(mediaMetadata.conductor);
+      }
+      if (mediaMetadata.discNumber != null) {
+        setDiscNumber(mediaMetadata.discNumber);
+      }
+      if (mediaMetadata.totalDiscCount != null) {
+        setTotalDiscCount(mediaMetadata.totalDiscCount);
+      }
+      if (mediaMetadata.genre != null) {
+        setGenre(mediaMetadata.genre);
+      }
+      if (mediaMetadata.compilation != null) {
+        setCompilation(mediaMetadata.compilation);
+      }
+      if (mediaMetadata.extras != null) {
+        setExtras(mediaMetadata.extras);
+      }
+
+      return this;
+    }
+
     /** Returns a new {@link MediaMetadata} instance with the current builder values. */
     public MediaMetadata build() {
       return new MediaMetadata(/* builder= */ this);