+++ library/common/src/test/java/com/google/android/exoplayer2/MediaMetadataTest.java
@@ -27,6 +27,9 @@ import org.junit.runner.RunWith;
 @RunWith(AndroidJUnit4.class)
 public class MediaMetadataTest {
 
+  private static final String EXTRAS_KEY = "exampleKey";
+  private static final String EXTRAS_VALUE = "exampleValue";
+
   @Test
   public void builder_minimal_correctDefaults() {
     MediaMetadata mediaMetadata = new MediaMetadata.Builder().build();
@@ -91,41 +94,62 @@ public class MediaMetadataTest {
   }
 
   @Test
-  public void roundTripViaBundle_yieldsEqualInstance() {
-    Bundle extras = new Bundle();
-    extras.putString("exampleKey", "exampleValue");
+  public void populate_populatesEveryField() {
+    MediaMetadata mediaMetadata = getFullyPopulatedMediaMetadata();
+    MediaMetadata populated = new MediaMetadata.Builder().populate(mediaMetadata).build();
 
-    MediaMetadata mediaMetadata =
-        new MediaMetadata.Builder()
-            .setTitle("title")
-            .setAlbumArtist("the artist")
-            .setMediaUri(Uri.parse("https://www.google.com"))
-            .setUserRating(new HeartRating(false))
-            .setOverallRating(new PercentageRating(87.4f))
-            .setArtworkData(
-                new byte[] {-88, 12, 3, 2, 124, -54, -33, 69}, MediaMetadata.PICTURE_TYPE_MEDIA)
-            .setTrackNumber(4)
-            .setTotalTrackCount(12)
-            .setFolderType(MediaMetadata.FOLDER_TYPE_PLAYLISTS)
-            .setIsPlayable(true)
-            .setRecordingYear(2000)
-            .setRecordingMonth(11)
-            .setRecordingDay(23)
-            .setReleaseYear(2001)
-            .setReleaseMonth(1)
-            .setReleaseDay(2)
-            .setComposer("Composer")
-            .setConductor("Conductor")
-            .setWriter("Writer")
-            .setDiscNumber(1)
-            .setTotalDiscCount(3)
-            .setGenre("Pop")
-            .setCompilation("Amazing songs.")
-            .setExtras(extras) // Extras is not implemented in MediaMetadata.equals(Object o).
-            .build();
+    // If this assertion fails, it's likely that a field is not being updated in
+    // MediaMetadata.Builder#populate(MediaMetadata).
+    assertThat(populated).isEqualTo(mediaMetadata);
+    assertThat(populated.extras.getString(EXTRAS_KEY)).isEqualTo(EXTRAS_VALUE);
+  }
+
+  @Test
+  public void roundTripViaBundle_yieldsEqualInstance() {
+    MediaMetadata mediaMetadata = getFullyPopulatedMediaMetadata();
 
     MediaMetadata fromBundle = MediaMetadata.CREATOR.fromBundle(mediaMetadata.toBundle());
     assertThat(fromBundle).isEqualTo(mediaMetadata);
-    assertThat(fromBundle.extras.getString("exampleKey")).isEqualTo("exampleValue");
+    // Extras is not implemented in MediaMetadata.equals(Object o).
+    assertThat(fromBundle.extras.getString(EXTRAS_KEY)).isEqualTo(EXTRAS_VALUE);
+  }
+
+  private static MediaMetadata getFullyPopulatedMediaMetadata() {
+    Bundle extras = new Bundle();
+    extras.putString(EXTRAS_KEY, EXTRAS_VALUE);
+
+    return new MediaMetadata.Builder()
+        .setTitle("title")
+        .setArtist("artist")
+        .setAlbumTitle("album title")
+        .setAlbumArtist("album artist")
+        .setDisplayTitle("display title")
+        .setSubtitle("subtitle")
+        .setDescription("description")
+        .setMediaUri(Uri.parse("https://www.google.com"))
+        .setUserRating(new HeartRating(false))
+        .setOverallRating(new PercentageRating(87.4f))
+        .setArtworkData(
+            new byte[] {-88, 12, 3, 2, 124, -54, -33, 69}, MediaMetadata.PICTURE_TYPE_MEDIA)
+        .setArtworkUri(Uri.parse("https://www.google.com"))
+        .setTrackNumber(4)
+        .setTotalTrackCount(12)
+        .setFolderType(MediaMetadata.FOLDER_TYPE_PLAYLISTS)
+        .setIsPlayable(true)
+        .setRecordingYear(2000)
+        .setRecordingMonth(11)
+        .setRecordingDay(23)
+        .setReleaseYear(2001)
+        .setReleaseMonth(1)
+        .setReleaseDay(2)
+        .setComposer("Composer")
+        .setConductor("Conductor")
+        .setWriter("Writer")
+        .setDiscNumber(1)
+        .setTotalDiscCount(3)
+        .setGenre("Pop")
+        .setCompilation("Amazing songs.")
+        .setExtras(extras)
+        .build();
   }
 }